hub/server-requirements.txt
python-engineio>=4.3.4
python-socketio>=5.8.0
numpy
eventlet>=0.33.3
greenlet>=2.0.2